pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "google-search-results>=2.4.2",
    "jupyter>=1.1.1",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "markdown>=3.8",
    "pygithub>=2.6.1",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "unstructured>=0.17.2",
]
